task6

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     static void divide(int toBeDivided) {
-        int result = toBeDivided/5;
+        double result = toBeDivided/5.0;
         System.out.println(result);
     }
 }
